+++ public/config.js
@@ -1,0 +1,10 @@
+module.exports = {
+    apiKey: "AIzaSyC1cldSI7eZjRZFNKEoKk5Yd35EcQ3qunc",
+    authDomain: "final-f637a.firebaseapp.com",
+    databaseURL: "https://final-f637a.firebaseio.com",
+    projectId: "final-f637a",
+    storageBucket: "final-f637a.appspot.com",
+    messagingSenderId: "752092036191",
+    appId: "1:752092036191:web:13cc92536e43e1227c2fea",
+    measurementId: "G-F9642WG9BS"
+}
