+++ public/index.html
@@ -5,9 +5,9 @@
     <title>UNC CS CompStone</title>
     <link rel='stylesheet' href='Style.css' type='text/css'>
     <script type="module" src="Render.js"></script>
-    <script src="engine/Cards.js"></script>
-    <script src="engine/cardGame.js"></script>
-    <script src="engine/AI.js"></script>
+    <script type="module" src="engine/Cards.js"></script>
+    <script type="module" src="engine/cardGame.js"></script>
+    <script type="module" src="engine/AI.js"></script>
 
 </head>
 <body>
